Handle blocked request better
This was spitting out an error to the console. For users that are
blocking this request with an ad blocker, probably best to just leave
this error out of the console.

For now, this will just add an attribute -- we could style using a
selector with the attribute, or handle some default content here in
the future.

Because of the CSS, a blocked ad will not alter display -- CSS styles
require the `loaded` class.