Add some user docs
This adds some docs that touchs on the ad placement types and the ad
placement themes. It actually uses the underlying `dist/ethicalads.js`,
so it's even possible to use this to review changes to the ad placement
styles in our PR builder :tada:

I absolutely phoned it in on the publisher sign up section. Let's add
some copy there.