Refactor some of the structure
Cleaning up some of the work that I blew through today, I wanted to move
everything that relates to handling of the Placement instances under the
Placement class. These were split between the class and a function for
iterating over the elements.

This was also cleanup to improve how we might drop in a viewport
detection method into the class -- the refactor makes this a nicer
pattern.

Very little logic changed, this is mostly restructuring the promise flow
to be cleaner and use state on the Placement instance.