Add viewport detection using the Verge module
- Uses the same detection method as RTD uses
- Adds a negligible ~1.5kb minified but not gzipped

One aspect of this that I don't like is that it uses setInterval
to check whether the ad is in the viewport.
This could be a minor performance hit although in most cases
it shouldn't be since it stops checking once it sees
the ad in the viewport.

In terms of rollout, it might be worth testing this
on some of our biggest publishers just to sanity check.