Output a UMD library instead of the default var library
This is another change that will require some testing, but I'm mostly
certain it does change anything about the current usage.

The default library target for Webpack is "var", which defines a global
variable, named `window.ethicalads` in this case. It, however, does not
export anything as a normal Node library would, so is not usable from
NPM directly.

This changes the library type to UMD, which in the Webpack config:

> "umd" ... exposes your library under all the module definitions, allowing it to work with CommonJS, AMD and as global variable

So, this seems to me to be only additive to the current configuration (global
variable). It's very worth some testing though.

This would also make it possible to package on NPM and install as a
library, though we'll need to tune up the automatic execution code as we
found this week on the RTD JS library work.